refactor: standardize theme imports across components and remove deprecated useThemeFromStore hook
- Updated all components to import useTheme from the new shared-native-ui package.
- Removed the obsolete useThemeFromStore hook and related themeStore.
- Adjusted imports for theme-related types and constants to align with the new package structure.

ref EL-125
