+++ apps/app-bible/src/app/App.tsx
@@ -14,12 +14,11 @@ import { QueryClientProvider } from '@tanstack/react-query';
 import { NavigationContainer } from '@react-navigation/native';
 import * as Linking from 'expo-linking';
 import { SystemBars } from 'react-native-edge-to-edge';
-import { useTheme } from '@/shared/hooks';
+import { useTheme, useThemeStore } from '@everylanguage/shared-native-ui';
 import { useLocalization } from '@/shared/hooks';
 import { useOnboardingStore } from '@/features/onboarding/store/onboardingStore';
 import { useAuthStore } from '@/shared/store/authStore';
 import { useLocalizationStore } from '@/shared/store/localizationStore';
-import { useThemeStore } from '@everylanguage/shared-native-ui';
 import { permissionsService } from '@/features/permissions/services/PermissionsService';
 import { powerSyncSystem } from '@/shared/services/powersync';
 import { authService } from '@/features/auth';
