+++ apps/app-bible/src/features/languages/components/VersionMenuButton.tsx
@@ -2,7 +2,7 @@ import React, { useMemo, useCallback } from 'react';
 import { TouchableOpacity } from 'react-native';
 import { Ionicons } from '@expo/vector-icons';
 import { MenuView, MenuAction } from '@react-native-menu/menu';
-import { useTheme } from '../../../shared/hooks/useThemeFromStore';
+import { useTheme } from '@/shared/hooks';
 import { logger } from '../../../shared/utils/logger';
 import type { AudioVersion, TextVersion } from '../types/entities';
 import { useLocalization } from '@/shared/hooks';
