+++ apps/app-bible/src/features/languages/screens/LanguageInfoScreen.tsx
@@ -11,7 +11,7 @@ import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { useNavigation } from '@react-navigation/native';
 import { Ionicons } from '@expo/vector-icons';
 import { MenuView, MenuAction } from '@react-native-menu/menu';
-import { useTheme } from '@/shared/hooks/useThemeFromStore';
+import { useTheme } from '@/shared/hooks';
 import { useLocalization } from '@/shared/hooks';
 import type {
   LanguageInfoScreenProps,
