+++ apps/app-bible/src/features/languages/screens/LanguageSearchScreen.tsx
@@ -9,7 +9,7 @@ import {
 import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { Ionicons } from '@expo/vector-icons';
 import { useNavigation } from '@react-navigation/native';
-import { useTheme } from '@/shared/hooks/useThemeFromStore';
+import { useTheme } from '@/shared/hooks';
 import { useLocalization } from '@/shared/hooks';
 import { useLanguageSearch } from '../hooks/useLanguageSearch';
 // import { useVersionsStore } from '../store/versionsStore';
