+++ apps/app-bible/src/shared/components/NoInternetModal.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { View, Text, StyleSheet, TouchableOpacity, Modal } from 'react-native';
 import { MaterialIcons } from '@expo/vector-icons';
 import { useTheme } from '@/shared/hooks';
-import { COLOR_VARIATIONS } from '@/shared/constants/theme';
+import { COLOR_VARIATIONS } from '@everylanguage/shared-native-ui';
 
 interface NoInternetModalProps {
   visible: boolean;
